Update ade20k_loader.py
Add assertion check, on line 27,  to see if split contains intended splits i.e. 'training' and 'validation'. Added a better way to read files of the intended split, on lines 45 and 46. Changed label path on line 57, which previously didn't point to the 'annotations' directory. Changed 'm.imread' to 'cv2.imread', on lines 59 and 62. Changed 'm.imresize' to 'cv2.resize', on lines 74 and 89. Removed assertion check on line 91.